5.1. Move render and common utility functions to seperate modules. Fix errors caused by that change
